Fix for <ul> scroll bar
Bug was: "When attempting to click and drag the scroll bar of the <ul>, the `.blur.fs` event would trigger and, triggering the `close.fs` event.  Added mouseenter and mouseleave to add active class to the trigger and conditional within `blur.fs` to detect active state before triggering `close.fs` event.